+++ src/FluentMigrator.Runner.Core/Processors/GenericProcessorBase.cs
@@ -16,36 +16,90 @@
 
 #endregion
 
+using System;
 using System.Data;
+using System.Data.Common;
+using System.Diagnostics;
+
+using JetBrains.Annotations;
 
 namespace FluentMigrator.Runner.Processors
 {
     public abstract class GenericProcessorBase : ProcessorBase
     {
+        [Obsolete]
         private readonly string _connectionString;
 
-        protected GenericProcessorBase(IDbConnection connection, IDbFactory factory
-                                       , IMigrationGenerator generator, IAnnouncer announcer, IMigrationProcessorOptions options)
+        [Obsolete]
+        protected GenericProcessorBase(
+            IDbConnection connection,
+            IDbFactory factory,
+            IMigrationGenerator generator,
+            IAnnouncer announcer,
+            [NotNull] IMigrationProcessorOptions options)
             : base(generator, announcer, options)
         {
-            Connection = connection;
+            DbProviderFactory = (factory as DbFactoryBase)?.Factory;
+
+            // Set the connection string, because it cannot be set by
+            // the base class (due to the missing information)
+            Options.ConnectionString = connection?.ConnectionString;
 
             // Prefetch connectionstring as after opening the security info could no longer be present
             // for instance on sql server
             _connectionString = connection?.ConnectionString;
 
             Factory = factory;
+
+            Connection = connection;
         }
 
-        public override string ConnectionString { get { return _connectionString; } }
+        protected GenericProcessorBase(
+            [CanBeNull] DbProviderFactory factory,
+            [NotNull] IMigrationGenerator generator,
+            [NotNull] IAnnouncer announcer,
+            [NotNull] ProcessorOptions options)
+            : base(generator, announcer, options)
+        {
+            DbProviderFactory = factory;
 
+#pragma warning disable 612
+            var legacyFactory = new DbFactoryWrapper(this);
+
+            // Prefetch connectionstring as after opening the security info could no longer be present
+            // for instance on sql server
+            _connectionString = options.ConnectionString;
+
+            Factory = legacyFactory;
+#pragma warning restore 612
+
+            if (factory != null)
+            {
+                Connection = factory.CreateConnection();
+                Debug.Assert(Connection != null, nameof(Connection) + " != null");
+                Connection.ConnectionString = options.ConnectionString;
+            }
+        }
+
+        [Obsolete]
+        public override string ConnectionString => _connectionString;
+
+        [CanBeNull]
         public IDbConnection Connection { get; protected set; }
+
+        [Obsolete]
+        [NotNull]
         public IDbFactory Factory { get; protected set; }
+
+        [CanBeNull]
         public IDbTransaction Transaction { get; protected set; }
+
+        [CanBeNull]
+        protected DbProviderFactory DbProviderFactory { get; }
 
         protected virtual void EnsureConnectionIsOpen()
         {
-            if (Connection.State != ConnectionState.Open)
+            if (Connection != null && Connection.State != ConnectionState.Open)
             {
                 Connection.Open();
             }
@@ -53,7 +107,7 @@ namespace FluentMigrator.Runner.Processors
 
         protected virtual void EnsureConnectionIsClosed()
         {
-            if (Connection.State != ConnectionState.Closed)
+            if (Connection != null && Connection.State != ConnectionState.Closed)
             {
                 Connection.Close();
             }
@@ -66,7 +120,8 @@ namespace FluentMigrator.Runner.Processors
             EnsureConnectionIsOpen();
 
             Announcer.Say("Beginning Transaction");
-            Transaction = Connection.BeginTransaction();
+
+            Transaction = Connection?.BeginTransaction();
         }
 
         public override void RollbackTransaction()
@@ -95,6 +150,60 @@ namespace FluentMigrator.Runner.Processors
         {
             RollbackTransaction();
             EnsureConnectionIsClosed();
+        }
+
+        protected virtual IDbCommand CreateCommand(string commandText)
+        {
+            IDbCommand result;
+            if (DbProviderFactory != null)
+            {
+                result = DbProviderFactory.CreateCommand();
+                Debug.Assert(result != null, nameof(result) + " != null");
+                result.Connection = Connection;
+                if (Transaction != null)
+                    result.Transaction = Transaction;
+                result.CommandText = commandText;
+            }
+            else
+            {
+#pragma warning disable 612
+                result = Factory.CreateCommand(commandText, Connection, Transaction, Options);
+#pragma warning restore 612
+            }
+
+            return result;
+        }
+
+        [Obsolete]
+        private class DbFactoryWrapper : IDbFactory
+        {
+            private readonly GenericProcessorBase _processor;
+
+            public DbFactoryWrapper(GenericProcessorBase processor)
+            {
+                _processor = processor;
+            }
+
+            /// <inheritdoc />
+            public IDbConnection CreateConnection(string connectionString)
+            {
+                Debug.Assert(_processor.DbProviderFactory != null, "_processor.DbProviderFactory != null");
+                var result = _processor.DbProviderFactory.CreateConnection();
+                Debug.Assert(result != null, nameof(result) + " != null");
+                result.ConnectionString = connectionString;
+                return result;
+            }
+
+            /// <inheritdoc />
+            [Obsolete]
+            public IDbCommand CreateCommand(
+                string commandText,
+                IDbConnection connection,
+                IDbTransaction transaction,
+                IMigrationProcessorOptions options)
+            {
+                return _processor.CreateCommand(commandText);
+            }
         }
     }
 }