+++ src/FluentMigrator.Runner.Db2/Processors/Db2/Db2ProcessorFactory.cs
@@ -20,6 +20,9 @@ using System;
 
 using FluentMigrator.Runner.Generators.DB2;
 
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
+
 namespace FluentMigrator.Runner.Processors.DB2
 {
     public class Db2ProcessorFactory : MigrationProcessorFactory
@@ -36,11 +39,22 @@ namespace FluentMigrator.Runner.Processors.DB2
             _serviceProvider = serviceProvider;
         }
 
+        [Obsolete]
         public override IMigrationProcessor Create(string connectionString, IAnnouncer announcer, IMigrationProcessorOptions options)
         {
             var factory = new Db2DbFactory(_serviceProvider);
             var connection = factory.CreateConnection(connectionString);
             return new Db2Processor(connection, new Db2Generator(new Db2Quoter()), announcer, options, factory);
+        }
+
+        /// <inheritdoc />
+        public override IMigrationProcessor Create()
+        {
+            var factory = new Db2DbFactory(_serviceProvider).Factory;
+            var options = _serviceProvider.GetRequiredService<IOptions<ProcessorOptions>>();
+            var announcer = _serviceProvider.GetRequiredService<IAnnouncer>();
+            var generator = new Db2Generator(new Db2Quoter());
+            return new Db2Processor(factory, generator, announcer, options);
         }
     }
 }