+++ src/FluentMigrator.Runner/LegacyExtensions.cs
@@ -24,6 +24,7 @@ using System.Reflection;
 using FluentMigrator.Infrastructure;
 using FluentMigrator.Runner.Initialization;
 using FluentMigrator.Runner.Initialization.AssemblyLoader;
+using FluentMigrator.Runner.Processors;
 using FluentMigrator.Runner.VersionTableInfo;
 
 using JetBrains.Annotations;
@@ -204,6 +205,26 @@ namespace FluentMigrator.Runner
         public static string GetName([NotNull] this IMigrationGenerator generator)
         {
             return generator.GetType().Name.Replace("Generator", string.Empty);
+        }
+
+        /// <summary>
+        /// Gets a <see cref="ProcessorOptions"/> instance for a given <see cref="IMigrationProcessorOptions"/> implementation
+        /// </summary>
+        /// <param name="options">The instance to get the <see cref="ProcessorOptions"/> for</param>
+        /// <param name="connectionString">The connection string</param>
+        /// <returns>The found/created <see cref="ProcessorOptions"/></returns>
+        internal static ProcessorOptions GetProcessorOptions(this IMigrationProcessorOptions options, string connectionString)
+        {
+            if (options == null)
+                return null;
+
+            return options as ProcessorOptions ?? new ProcessorOptions()
+            {
+                ConnectionString = connectionString,
+                PreviewOnly = options.PreviewOnly,
+                ProviderSwitches = options.ProviderSwitches,
+                Timeout = options.Timeout == null ? null : (TimeSpan?) TimeSpan.FromSeconds(options.Timeout.Value),
+            };
         }
     }
 }