+++ src/FluentMigrator.Runner/Processors/ConnectionlessProcessorFactory.cs
@@ -14,10 +14,14 @@
 // limitations under the License.
 #endregion
 
+using System;
+
 using FluentMigrator.Runner.Generators;
 using FluentMigrator.Runner.Initialization;
 
 using JetBrains.Annotations;
+
+using Microsoft.Extensions.Options;
 
 namespace FluentMigrator.Runner.Processors
 {
@@ -30,24 +34,54 @@ namespace FluentMigrator.Runner.Processors
         private readonly IMigrationGenerator _generator;
 
         [NotNull]
-        private readonly IRunnerContext _runnerContext;
+        private readonly string _databaseId;
+
+        [NotNull]
+        private readonly ProcessorOptions _options;
+
+        [NotNull]
+        private readonly IAnnouncer _announcer;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ConnectionlessProcessorFactory"/> class.
         /// </summary>
         /// <param name="generatorAccessor">The accessor to get the migration generator to use</param>
         /// <param name="runnerContext">The runner context</param>
-        public ConnectionlessProcessorFactory([NotNull] IMigrationGeneratorAccessor generatorAccessor, [NotNull] IRunnerContext runnerContext)
+        [Obsolete]
+        public ConnectionlessProcessorFactory(
+            [NotNull] IMigrationGeneratorAccessor generatorAccessor,
+            [NotNull] IRunnerContext runnerContext)
         {
             _generator = generatorAccessor.ActiveGenerator;
-            _runnerContext = runnerContext;
+            _databaseId = runnerContext.Database;
+            _announcer = runnerContext.Announcer;
+            _options = new ProcessorOptions(runnerContext);
             Name = _generator.GetName();
         }
 
+        public ConnectionlessProcessorFactory(
+            [NotNull] IMigrationGeneratorAccessor generatorAccessor,
+            [NotNull] IOptions<ProcessorOptions> options,
+            [NotNull] IAnnouncer announcer)
+        {
+            _generator = generatorAccessor.ActiveGenerator;
+            _options = options.Value;
+            _announcer = announcer;
+            Name = _databaseId = _generator.GetName();
+        }
+
         /// <inheritdoc />
+        [Obsolete]
         public IMigrationProcessor Create(string connectionString, IAnnouncer announcer, IMigrationProcessorOptions options)
         {
-            return new ConnectionlessProcessor(_generator, _runnerContext, options);
+            var processorOptions = options.GetProcessorOptions(connectionString);
+            return new ConnectionlessProcessor(_databaseId, _generator, _announcer, processorOptions);
+        }
+
+        /// <inheritdoc />
+        public IMigrationProcessor Create()
+        {
+            return new ConnectionlessProcessor(_databaseId, _generator, _announcer, _options);
         }
 
         /// <inheritdoc />