+++ test/FluentMigrator.Tests/Helpers/Db2TestTable.cs
@@ -102,10 +102,7 @@ namespace FluentMigrator.Tests.Helpers
             var columns = string.Join(", ", columnDefinitions);
             sb.AppendFormat("CREATE TABLE {0} ({1})", NameWithSchema, columns);
 
-            using (var command = Factory.CreateCommand(sb.ToString(), Connection, Transaction, Processor.Options))
-            {
-                command.ExecuteNonQuery();
-            }
+            Processor.Execute(sb.ToString());
         }
 
         public void Dispose()
@@ -116,20 +113,12 @@ namespace FluentMigrator.Tests.Helpers
         public void Drop()
         {
             var tableCommand = string.Format("DROP TABLE {0}", NameWithSchema);
-
-            using (var command = Factory.CreateCommand(tableCommand, Connection, Transaction, Processor.Options))
-            {
-                command.ExecuteNonQuery();
-            }
+            Processor.Execute(tableCommand);
 
             if (!string.IsNullOrEmpty(_schema))
             {
                 var schemaCommand = string.Format("DROP SCHEMA {0} RESTRICT", _quoter.QuoteSchemaName(_schema));
-
-                using (var commandToo = Factory.CreateCommand(schemaCommand, Connection, Transaction, Processor.Options))
-                {
-                    commandToo.ExecuteNonQuery();
-                }
+                Processor.Execute(schemaCommand);
             }
         }
 
@@ -141,10 +130,7 @@ namespace FluentMigrator.Tests.Helpers
                 _quoter.QuoteColumnName(column)
                 );
 
-            using (var command = Factory.CreateCommand(query, Connection, Transaction, Processor.Options))
-            {
-                command.ExecuteNonQuery();
-            }
+            Processor.Execute(query);
         }
 
         public void WithUniqueConstraintOn(string column, string name)
@@ -157,10 +143,7 @@ namespace FluentMigrator.Tests.Helpers
                 _quoter.QuoteColumnName(column)
             );
 
-            using (var command = Factory.CreateCommand(query, Connection, Transaction, Processor.Options))
-            {
-                command.ExecuteNonQuery();
-            }
+            Processor.Execute(query);
         }
 
         public Db2Processor Processor { get; }