+++ test/FluentMigrator.Tests/Integration/Processors/Oracle/OracleProcessorTestsBase.cs
@@ -119,13 +119,9 @@ namespace FluentMigrator.Tests.Integration.Processors.Oracle {
         public void TestQuery()
         {
             string sql = "SELECT SYSDATE FROM " + Quoter.QuoteTableName("DUAL");
-            using (var command = Factory.CreateCommand(sql, Processor.Connection, Processor.Transaction, Processor.Options))
-            using (var reader = command.ExecuteReader())
-            {
-                var ds = reader.ReadDataSet();
-                Assert.Greater(ds.Tables.Count, 0);
-                Assert.Greater(ds.Tables[0].Columns.Count, 0);
-            }
+            var ds = Processor.Read(sql);
+            Assert.Greater(ds.Tables.Count, 0);
+            Assert.Greater(ds.Tables[0].Columns.Count, 0);
         }
     }
 }